% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{exploreDataFrameAddin}
\alias{exploreDataFrameAddin}
\title{Insert statements for exploratory data analysis of a data frame}
\usage{
exploreDataFrameAddin()
}
\value{
Statements for exploring a data frame, including head,
         tail, colnames, column types, and dimensions
}
\description{
Insert statements for exploratory data analysis of a data frame
}

