% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compbio-addins.R
\name{exploreExpressionSetAddin}
\alias{exploreExpressionSetAddin}
\title{Explore microarray data (ExpressionSet)}
\usage{
exploreExpressionSetAddin()
}
\description{
Explore microarray data (ExpressionSet)
}

