% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{mapAddin}
\alias{mapAddin}
\title{Insert anonymous function map statement}
\usage{
mapAddin()

mapAddin()
}
\value{
Anonymous function map statement
}
\description{
Insert anonymous function map statement

@export mapUnlistAddin
}

