% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd-addins.R
\name{matrix2latex}
\alias{matrix2latex}
\title{Convert matrix output to latex}
\usage{
matrix2latex(matr)
}
\arguments{
\item{matr}{Numeric matrix}
}
\value{
LaTeX typesetted matrix
}
\description{
Convert matrix output to latex
}
\examples{
{
matr <- matrix(data = c(1, 0, 0,
                        0, 1, 0,
                        0, 0, 1), nrow = 3, ncol = 3)
#matrix2latex(matr)
}
}

