% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compbio-addins.R
\name{microarrayExpressAddin}
\alias{microarrayExpressAddin}
\title{Load GEO entry from NCBI and explore ExpressionSet
 varLabels(featureData(ma[[1]]))}
\usage{
microarrayExpressAddin()
}
\description{
Load GEO entry from NCBI and explore ExpressionSet
 varLabels(featureData(ma[[1]]))
}

