% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{summarizeCrvAddin}
\alias{summarizeCrvAddin}
\title{Summarize and plot a histogram of a continuous random variable}
\usage{
summarizeCrvAddin()
}
\description{
Summarize and plot a histogram of a continuous random variable
}

